+++ packages/native/src/lib.rs
@@ -1,7 +1,7 @@
 #![deny(clippy::all)]
 
-mod automation;
 mod appshot;
+mod automation;
 mod color;
 mod custom_elements;
 mod diff;
@@ -18,7 +18,7 @@ mod theme;
 #[cfg(all(feature = "test-support", target_os = "macos"))]
 mod test_renderer;
 
-pub use element_tree::*;
 pub use appshot::*;
+pub use element_tree::*;
 pub use renderer::*;
 pub use style::*;